all: Rename flask.py file to avoid name conflicts
Running the apps directly with python was causing ImportErrors, as:
__init__.py would import module.flask.py
module.flask would import backend_common
backend_common would try to import something from flask, and would reach module.flask.py not the Flask library.
