+++ Makefile
@@ -201,7 +201,7 @@ develop-run-BACKEND: build-certs nix require-APP
 	APP_URL=https://$(APP_DEV_HOST):$(APP_DEV_PORT_$(APP)) \
 	CORS_ORIGINS="*" \
 		nix-shell nix/default.nix -A $(APP) \
-			--run "gunicorn $(APP_PYTHON).flask:app --bind '$(APP_DEV_HOST):$(APP_DEV_PORT_$(APP))' --ca-certs=$$PWD/tmp/ca.crt --certfile=$$PWD/tmp/server.crt --keyfile=$$PWD/tmp/server.key --workers 1 --timeout 3600 --reload --log-file -"
+			--run "gunicorn $(APP_PYTHON).flask_control:app --bind '$(APP_DEV_HOST):$(APP_DEV_PORT_$(APP))' --ca-certs=$$PWD/tmp/ca.crt --certfile=$$PWD/tmp/server.crt --keyfile=$$PWD/tmp/server.key --workers 1 --timeout 3600 --reload --log-file -"
 
 develop-run-FRONTEND: build-certs nix require-APP
 	nix-shell nix/default.nix --pure -A $(APP) \
@@ -230,7 +230,7 @@ develop-flask-shell: nix require-APP
 	DEBUG=true \
 	CACHE_TYPE=filesystem \
 	CACHE_DIR=$$PWD/src/$(APP)/cache \
-	FLASK_APP=$(APP).flask \
+	FLASK_APP=$(APP).flask_control \
 	APP_SETTINGS=$$PWD/src/$(APP)/settings.py \
 		nix-shell nix/default.nix -A $(APP) \
 		--run "flask $(FLASK_CMD)"
