+++ src/releng_mapper/releng_mapper/__init__.py
@@ -2,7 +2,7 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-from releng_mapper.flask import app
+from releng_mapper.flask_control import app
 
 
 if __name__ == "__main__":
