+++ src/releng_tooltool/releng_tooltool/__init__.py
@@ -2,7 +2,7 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-from releng_tooltool.flask import app
+from releng_tooltool.flask_control import app
 
 
 if __name__ == "__main__":
