+++ src/releng_treestatus/releng_treestatus/__init__.py
@@ -2,7 +2,7 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-from releng_treestatus.flask import app
+from releng_treestatus.flask_control import app
 
 
 if __name__ == "__main__":
